requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.4
python-dotenv==1.0.1
anthropic==0.43.0
requests==2.32.3
pandas==2.2.3
ta==0.11.0
psycopg2-binary==2.9.10
schedule==1.2.2
numpy==1.26.4
PyJWT==2.9.0
cryptography==43.0.3
